+++ ui/src/features/project/pipelines/pipelines.tsx
@@ -38,7 +38,7 @@ import {
 import { FreightTimeline } from './freight/freight-timeline';
 import { Graph } from './graph/graph';
 import { GraphFilters } from './graph-filters';
-import { Images } from './images';
+import { Images } from './image-history/images';
 import { PipelineListView } from './list/list-view';
 import { Promote } from './promotion/promote';
 import { Promotion } from './promotion/promotion';
